+++ packages/app-builder-lib/src/configuration.ts
@@ -131,6 +131,11 @@ export interface Configuration extends PlatformSpecificBuildOptions {
    * @default true
    */
   readonly npmRebuild?: boolean
+  /**
+   * Use `legacy` app-builder binary for installing native dependencies, or `@electron/rebuild` in `sequential` or `parallel` compilation modes.
+   * @default sequential
+   */
+  readonly nativeRebuilder?: "legacy" | "sequential" | "parallel" | null
 
   /**
    * The build number. Maps to the `--iteration` flag for builds using FPM on Linux.
@@ -216,6 +221,11 @@ export interface Configuration extends PlatformSpecificBuildOptions {
   readonly beforePack?: ((context: BeforePackContext) => Promise<any> | any) | string | null
 
   /**
+   * The function (or path to file or module id) to be [run after the prebuilt Electron binary has been extracted to the output directory](#afterextract)
+   */
+  readonly afterExtract?: ((context: AfterExtractContext) => Promise<any> | any) | string | null
+
+  /**
    * The function (or path to file or module id) to be [run after pack](#afterpack) (but before pack into distributable format and sign).
    */
   readonly afterPack?: ((context: AfterPackContext) => Promise<any> | any) | string | null
@@ -274,6 +284,12 @@ export interface Configuration extends PlatformSpecificBuildOptions {
    * @default true
    */
   readonly removePackageKeywords?: boolean
+
+  /**
+   * Whether to disable sanity check asar package (useful for custom electron forks that implement their own encrypted integrity validation)
+   * @default false
+   */
+  readonly disableSanityCheckAsar?: boolean
 }
 
 interface PackContext {
@@ -286,6 +302,7 @@ interface PackContext {
 }
 export type AfterPackContext = PackContext
 export type BeforePackContext = PackContext
+export type AfterExtractContext = PackContext
 
 export interface MetadataDirectories {
   /**