+++ packages/builder-util/src/util.ts
@@ -17,12 +17,15 @@ if (process.env.JEST_WORKER_ID == null) {
 
 export { safeStringifyJson } from "builder-util-runtime"
 export { TmpDir } from "temp-file"
-export { log, debug } from "./log"
+export * from "./log"
 export { Arch, getArchCliNames, toLinuxArchString, getArchSuffix, ArchType, archFromString, defaultArchFromString } from "./arch"
 export { AsyncTaskManager } from "./asyncTaskManager"
 export { DebugLogger } from "./DebugLogger"
+export { httpExecutor, NodeHttpExecutor } from "./nodeHttpExecutor"
+export * from "./promise"
+export * from "./arch"
 
-export { copyFile, exists } from "./fs"
+export * from "./fs"
 export { asArray } from "builder-util-runtime"
 
 export { deepAssign } from "./deepAssign"
@@ -404,14 +407,14 @@ export async function executeAppBuilder(
   }
 }
 
-export async function retry<T>(task: () => Promise<T>, retriesLeft: number, interval: number, backoff = 0, attempt = 0): Promise<T> {
+export async function retry<T>(task: () => Promise<T>, retryCount: number, interval: number, backoff = 0, attempt = 0, shouldRetry?: (e: any) => boolean): Promise<T> {
   try {
     return await task()
   } catch (error: any) {
-    log.info(`Above command failed, retrying ${retriesLeft} more times`)
-    if (retriesLeft > 0) {
+    log.info(`Above command failed, retrying ${retryCount} more times`)
+    if ((shouldRetry?.(error) ?? true) && retryCount > 0) {
       await new Promise(resolve => setTimeout(resolve, interval + backoff * attempt))
-      return await retry(task, retriesLeft - 1, interval, backoff, attempt + 1)
+      return await retry(task, retryCount - 1, interval, backoff, attempt + 1, shouldRetry)
     } else {
       throw error
     }