+++ packages/electron-builder/src/index.ts
@@ -1,5 +1,6 @@
-export { getArchSuffix, Arch, archFromString } from "builder-util"
+export { getArchSuffix, Arch, archFromString, log } from "builder-util"
 export { build, CliOptions, createTargets } from "./builder"
+export { publish, publishArtifactsWithOptions } from "./publish"
 export {
   TargetConfiguration,
   Platform,
@@ -63,6 +64,9 @@ export {
   PublishManager,
   PublishOptions,
   ProgressInfo,
+  MacPackager,
+  WinPackager,
+  LinuxPackager,
 } from "app-builder-lib"
 export { buildForge, ForgeOptions } from "app-builder-lib"
 export { CancellationToken } from "builder-util-runtime"